reworking itnernal protocol, remaking the lora driver on both ends to be non blocking and adding extra funcitonality. ready to test server side to do latency measurements

+++ client/boot.py
@@ -28,62 +28,46 @@ led_builtin = machine.Pin(25, machine.Pin.OUT)
 animationutils.current_animation_flag = 'none' #shared flags between all threads and utils.py
 animationutils.last_animation_flag = 'none'
 
-
-
-
+reference = {
+    0:'off',
+    1:'test',
+    2:'spotlight',
+    3:'nature',
+    4:'dystopia',
+    5:'irl'
+}
 
 def listen_to_host(): #this is the second core functionality
     while 1:
         received_msg = rylr.receive()
         if not received_msg == None:
-            print(received_msg)
-            data_parsed = received_msg.data.decode("ascii").split()
-            #check received via imitating match-case; micropython does not have it :(
+            #print(received_msg)
 
-            #if not an animation command, handle command
-            if data_parsed[0] == 'pyexec':
-                command = received_msg.data.decode("ascii")[7:]
-                try:
-                    exec(command)
-                except Exception as e:
-                    print(e)
+            #if not 2 bytes, so is not an animation command, handle command
+            if len(received_msg.data) !=2:
+                data_parsed = received_msg.data.decode("ascii").split()
+                if data_parsed[0] == 'pyexec':
+                    command = received_msg.data.decode("ascii")[7:]
+                    try:
+                        exec(command)
+                    except Exception as e:
+                        print(e)
 
-            elif data_parsed[0] == 'nuke':#failsafe/remote debugging - deletes boot py in case it blocks serial/fails something else
-                print('got nuke, dleteing boot and restetting machine...')
-                os.remove('boot.py')
-                machine.reset()
+                elif data_parsed[0] == 'nuke':#failsafe/remote debugging - deletes boot py in case it blocks serial/fails something else
+                    print('got nuke, dleteing boot and restetting machine...')
+                    os.remove('boot.py')
+                    machine.reset()
 
-            #if an anitmaiton command, change the current_animation flag
-            elif data_parsed[0] == 'presetoff':
+            else:
+                #if an anitmaiton command, change the current_animation flag
+                msg_unpacked = utils.unpack_octal(received_msg.data) 
+
                 lock.acquire()
-                animationutils.current_animation_flag = 'off'
-                lock.release()     
-            
-            elif data_parsed[0] == 'presettest':
-                lock.acquire()
-                animationutils.current_animation_flag = 'test'
+                animationutils.current_animation_flag = reference[msg_unpacked[-1]]
                 lock.release()
-            
-            elif data_parsed[0] == 'presetspotlight':
-                lock.acquire()
-                animationutils.current_animation_flag = 'spotlight'
-                lock.release()     
-            
-            elif data_parsed[0] == 'presetnature':
-                lock.acquire()
-                animationutils.current_animation_flag = 'nature'
-                lock.release()
-                
-            elif data_parsed[0] == 'presetdystopia':
-                lock.acquire()
-                animationutils.current_animation_flag = 'dystopia'   
-                lock.release()
-                
-            elif data_parsed[0] == 'presetirl':
-                lock.acquire()
-                animationutils.current_animation_flag = 'irl' 
-                lock.release()
-        time.sleep_ms(1)
+                #print(msg_unpacked[-1])
+
+        time.sleep_us(10)
 
 
 print('> Setting up LoRa')
@@ -103,23 +87,32 @@ while True: #start actively tracking animations
         if animationutils.last_animation_flag != cur_anim_local:   
             if animationutils.last_animation_flag != 'none':
                 animationutils.AnimationInstance.kill_all() 
-            if cur_anim_local == 'nature':
-                nature_animation.reset()
-                nature_animation.play() 
-            elif cur_anim_local == 'off':
-                this_servo.set(0)
-                this_jewel.set(0)
-            elif cur_anim_local == 'irl':
-                irl_animation.reset()
-                irl_animation.play()
+            if cur_anim_local == 'test':
+                servo_test.reset(); jewel_test.reset()
+                servo_test.play(); jewel_test.play()
+
+            elif cur_anim_local == 'spotlight':
+                servo_spotlight.reset(); jewel_spotlight.reset()
+                servo_spotlight.play(); jewel_spotlight.play()
+
+            elif cur_anim_local == 'nature':
+                servo_nature.reset(); jewel_nature.reset()
+                servo_nature.play(); jewel_nature.play()
+
             elif cur_anim_local == 'dystopia':
-                irl_animation.reset()
-                irl_animation.play()
-                nature_animation.reset()
-                nature_animation.play() 
+                servo_dystopia.reset(); jewel_dystopia.reset()
+                servo_dystopia.play(); jewel_dystopia.play()
+
+            elif cur_anim_local == 'irl':
+                servo_irl.reset(); jewel_irl.reset()
+                servo_irl.play(); jewel_irl.play()
+
+            elif cur_anim_local == 'off':
+                this_servo.set(0) 
+                this_jewel.set(0)
     
         animationutils.last_animation_flag = cur_anim_local  
-        time.sleep_ms(1)
+        time.sleep_us(10)
     except Exception as e:
         print_exception(e)
         break
